test(REST API): Tests moved to db-access/api; Mocks & FW moved to testing
