+++ apps/site/__tests__/courses/[courseSlug]/[lessonSlug]/index.spec.tsx
@@ -3,7 +3,7 @@ import { database } from "@self-learning/database";
 import { getServerSession } from "next-auth";
 import { createMockContext } from "../../../context-utils";
 import { compileMarkdown } from "@self-learning/markdown";
-import { createCourseMock, createLessonMock } from "../../../mock-data-utils";
+import { createCourseMock, createLessonMock } from "@self-learning/util/testing";
 
 // Mocks getServerSession of withAuth procedure to mock User object
 // ESM support & default mock required by NextAuth
