+++ libs/util/testing/src/index.ts
@@ -2,3 +2,4 @@ export * from "./lib/example-data-generator";
 export * from "./lib/database-helpers";
 export * from "./lib/mocks";
 export * from "./lib/testing-commands";
+export * from "./lib/openapi-utils";
