+++ libs/util/testing/src/lib/openapi-utils.ts
